권한이 없으면 해당페이지 접속할수 없는 미들웨어 만들기

+++ 2022풀스택부트캠프/YelpCamp/routes/campgrounds.js
@@ -22,8 +22,8 @@ router.get(
   "/",
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const campgrounds = await Campground.find({});
-    res.render("campgrounds/index", { campgrounds });
+    const campground = await Campground.find({});
+    res.render("campgrounds/index", { campground });
   })
 );
 
@@ -70,10 +70,17 @@ router.get(
   "/:id/edit",
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
     if (!campground) {
       req.flash("error", "Cannot find that campground!");
       return res.redirect("/campgrounds");
+    }
+
+    // 사용자가 아니면 접속할수 없다.
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("Error", "사용자가 아닙니다!!!!");
+      return res.redirect(`/campgrounds/${id}`);
     }
     res.render("campgrounds/edit", { campground });
   })
@@ -85,7 +92,14 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+
+    // 사용자가 아니면 접속할수 없다.
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("Error", "사용자가 아닙니다!!!!");
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
     req.flash("success", "Successfully update campground");
@@ -98,6 +112,11 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("Error", "사용자가 아닙니다!!!!");
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash("success", "Successfully Deleted Campground");
     res.redirect("/campgrounds");
